Fixes #330 current NULL for mysqli
