+++ src/Adapter/Driver/Mysqli/Result.php
@@ -52,9 +52,9 @@ class Result implements
     protected $nextComplete = false;
 
     /**
-     * @var bool
+     * @var mixed
      */
-    protected $currentData = false;
+    protected $currentData = null;
 
     /**
      *
